refacto redux file's
